requirements.txt.txt
kafka-python
google-cloud-bigquery
pandas
